+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - dashboard panel editor: fix edit
 - dashboard panel img and CSS option 
 
+
 # 0.28.1
 - fix dashboard id